backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.4.0
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.23
pydantic==2.5.3
python-dotenv==1.0.0
podcastfy==0.4.1
audioop-lts==0.2.2
